deploy latest changes to Estimator (putting on pause indefinitely to focus on other aspects)
The estimator calculator has been laid out almost to completion. Some aesthetic changes still needed. Functionality wise, the form control captures the different aspects that I initially wanted to cover. The equation that performs the calculation is still needs development. Currently it calculates based of varies components that are dependent upon one another, I want to re-factor it so that time is the main dependency. Also the equation needs to capture all the aspects present in the form control, currently only uses the room size to calculate. This will be extensive and thus being put on hold, to hopefully add new parts to the site. :sob: